+++ src/commands/General/Chat Bot Info/stats.js
@@ -1,6 +1,5 @@
-const { Command, version: klasaVersion } = require('klasa');
+const { Command, version: klasaVersion, Timestamp } = require('klasa');
 const { version: discordVersion } = require('discord.js');
-const moment = require('moment');
 
 module.exports = class extends Command {
 
@@ -14,7 +13,7 @@ module.exports = class extends Command {
 	async run(msg) {
 		return msg.sendCode('asciidoc', msg.language.get('COMMAND_STATS',
 			(process.memoryUsage().heapUsed / 1024 / 1024).toFixed(2),
-			moment(Date.now() - (process.uptime() * 1000)).toNow(true),
+			Timestamp.toNow(Date.now() - (process.uptime() * 1000), true),
 			this.client.users.size.toLocaleString(),
 			this.client.guilds.size.toLocaleString(),
 			this.client.channels.size.toLocaleString(),