+++ typings/index.d.ts
@@ -701,7 +701,7 @@ declare module 'klasa' {
 		public constructor(options: KlasaConsoleConfig);
 		public readonly stdout: NodeJS.WritableStream;
 		public readonly stderr: NodeJS.WritableStream;
-		public timestaamps: boolean | string;
+		public template?: Timestamp;
 		public useColors: boolean;
 		public colors: KlasaConsoleColorsOption;
 
@@ -719,6 +719,8 @@ declare module 'klasa' {
 		public static flatten(data: any, useColors: boolean): string;
 	}
 
+	export { KlasaConsole as Console };
+
 	export class Stopwatch {
 		public constructor(digits?: number);
 		private _start: number;
@@ -735,7 +737,19 @@ declare module 'klasa' {
 		public toString(): string;
 	}
 
-	export { KlasaConsole as Console };
+	export class Timestamp {
+		public constructor(pattern: string);
+		public pattern: string;
+		private _template: TimestampObject[];
+
+		public display(time: Date | number | string): string;
+		public edit(pattern: string): this;
+
+		private _parse(type: string, time: Date): string;
+		private _patch(pattern: string): void;
+
+		public static toNow(earlier: Date | number | string, showIn?: boolean): string;
+	}
 
 	// Structures
 	export class Piece {
@@ -1246,6 +1260,11 @@ declare module 'klasa' {
 		debug?: boolean;
 		verbose?: boolean;
 		wtf?: boolean;
+	};
+
+	export type TimestampObject = {
+		type: string;
+		content?: string;
 	};
 
 	export type PermissionLevel = {